changed data types

+++ Maths/mod_inverse_of_n.cpp
@@ -1,10 +1,10 @@
-long long mod_Inverse(long long a, long long uMOD = (long long)1e9+7ll) { 
-    long long m0 = uMOD; 
-    long long y = 0ll, x = 1ll; 
+int64_t mod_Inverse(int64_t a, int64_t uMOD = (int64_t)1e9+7ll) { 
+    int64_t m0 = uMOD; 
+    int64_t y = 0ll, x = 1ll; 
     if (uMOD == 1ll) return 0ll; 
     while (a > 1ll) { 
-        long long q = a / uMOD; 
-        long long t = uMOD; 
+        int64_t q = a / uMOD; 
+        int64_t t = uMOD; 
         uMOD = a % uMOD, a = t; t = y; 
         y = x - q * y; 
         x = t;
